added print statements to help debug. Also made it so that only 10 images are loaded (and displayed, press a key to go through them). I fixed the code in chooseAttribute, so now we don't crash. Now we just need to work on where we should catch things when the max information gain is 0 for the rest of the remaining features.

+++ trunk/decisionTree.h
@@ -38,7 +38,7 @@ private:
 	bool sameClassification(const std::vector<CClassifier::HaarOutput*> &examples);
 	bool isAttribsEmpty(const std::vector<bool> &attribs);
 	void setMajorityValues(const std::vector<CClassifier::HaarOutput*> &examples);
-	double EntropyFunc(double p);
+	long double EntropyFunc(long double p);
 
 
 	std::vector<DecisionTree*> children;
